strace.c: move argc/nprocs check to the tracing_backend_init routine
It is deemed that for some tracing backends it is possible to run strace
without providing either program to run or PIDs to attach (attaching to
already running GDB server with existing inferiors is an example). Thus,
let's move this check to tracing_backend_init.

* strace.c (tracing_backend_init): Add (!nprocs && !argc) check.
(init): Remove (!nprocs && !argc) check, update error message for (argc
< 0) check.
* tests/options-syntax.test: Update expected output.