configure.ac: use a separate compiler for AArch32/ARMv7 EABI on AArch64
As gcc on AArch64 doesn't allow compiling compat ABI binaries, we have
to (find and) use a separate compiler in order to get proper compat
personality support on AArch64.

* configure.ac (case "$host_cpu" in) <aarch64*>: Inform about
architecture detection results early. Try to find a compiler for ARM
EABI based on known names for it in various distributions.